Prop adjustments to make it one view and edit space
