+++ src/app/view/event/[id]/components/eventTabContent.tsx
@@ -1,8 +1,8 @@
 import { Tabs, TabsContent } from "@/components/ui/tabs";
 import EventDisplay from "./eventDisplay";
 import ContactSection from "./contactSection";
-import DocumentList from "./documentList";
 import DocumentTable from "./documentTable";
+import { IEventFrontend, ITempEventData } from "./event";
 
 interface IDocument {
   name: string;
@@ -12,6 +12,7 @@ interface IDocument {
 interface EventTabContentProps {
   eventId: string;
   activeTab: string;
+  eventData: (IEventFrontend & ITempEventData) | null;
   eventDetails: string;
   eventStatus: "Upcoming" | "Ongoing" | "Completed" | "Cancelled";
   vendorList: string;
@@ -19,15 +20,20 @@ interface EventTabContentProps {
   name?: string;
   email: string;
   phone: string;
+  editCache: (IEventFrontend & ITempEventData) | null;
   isEditing: boolean;
   isAdmin: boolean;
   onUpdateField: (field: string, value: any) => void;
+  onSave: () => void;
   onRemoveDocument: (index: number) => void;
+  setEventData: React.Dispatch<React.SetStateAction<(IEventFrontend & ITempEventData) | null>>;
+  setEditCache: React.Dispatch<React.SetStateAction<(IEventFrontend & ITempEventData) | null>>;
 }
 
 export default function EventTabContent({
   eventId,
   activeTab,
+  eventData,
   eventDetails,
   eventStatus,
   vendorList,
@@ -35,9 +41,13 @@ export default function EventTabContent({
   name,
   email,
   phone,
+  editCache,
   isEditing,
   isAdmin,
   onUpdateField,
+  onSave,
+  setEventData,
+  setEditCache,
   onRemoveDocument,
 }: EventTabContentProps) {
   return (
@@ -48,11 +58,16 @@ export default function EventTabContent({
           <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
             {/* Left Column - Event Details */}
             <EventDisplay
+              eventData={eventData}
               eventDetails={eventDetails}
               vendorList={vendorList}
+              editCache={editCache}
               isEditing={isEditing}
               isAdmin={isAdmin}
               onUpdateField={onUpdateField}
+              onSave={onSave}
+              setEventData={setEventData}
+              setEditCache={setEditCache}
             />
 
             {/* Right Column - Contact & Documents */}
